requirements.txt
numpy
sympy
django-cors-headers
djangorestframework

